Fixed #1131 - Race condition between `SuspendNotifications()` and `.ResumeNotifications()` (#1132)
* Fixed that `ObservableCache.SuspendNotifications()` uses a `lock` to synchronize updates to internal state for tracking notification suspensions, but `.ResumeNotifications()` did not use any synchronization when updating that same state, allowing for corruption of state if the two methods happen to overlap, concurrently.

Also fixed that the test for this behavior was intermittent, specifically that it tended to falsely pass when running in parallel with many other tests. The test has now been moved to its own `IntegrationTests` fixture, to guarantee it is never parallelized with other tests.

Resolves #1131.

* Deliver resumed notifications off the lock, closing the #1131 race without holding it

ResumeNotifications() was wrapped in lock(_locker), which via lock reentrancy
held the cache lock across the DeliveryQueue drain, delivering accumulated
changes to subscribers while the lock was held. Restore off-lock delivery and
close the suspend/resume state-divergence race a different way:

- SuspendNotifications() no longer wraps the resume callback in the cache lock;
  queued changes drain outside the lock, as the DeliveryQueue intends.
- ResumeNotifications() re-checks the suspend count under the lock before
  emitting the resume signal, so a concurrent SuspendNotifications() in the
  decrement/emit window wins and the count and the subject cannot diverge.
- SuspensionTracker.SuspendNotifications() gates the suspended signal on the
  subject's own value rather than the count, keeping it monotonic while a
  resume's signal is still in flight.

Tests:
- Add StaleResumeSignalIsSuppressedByConcurrentReSuspend: deterministic proof
  that a connection made during a racing re-suspend does not activate on a
  stale resume signal (fails without the count re-check).
- Add ResumeDeliversPendingChangesWithoutHoldingTheLock: proves a blocked
  subscriber does not stall a concurrent lock-requiring operation during
  resume delivery.
- Remove ResumeSignalUnderLockPreventsStaleSnapshotFromReSuspend: redundant
  coverage that only passed by timing out a deadlock.

* Adjusted SuspendNotificationsFixture.UnitTests, for consistent styling.

* Moved an additional test exercising concurrency, that got missed, from SuspendNotificationsFixture.UnitTests to SuspendNotificationsFixture.IntegrationTests.

* Close the suspend/resume race by emitting the resume signal in a single lock acquisition

ResumeNotifications() was acquiring the lock twice: once to decrement the suspend
count and queue the accumulated changes, and again to emit the resume signal. The
gap between those two acquisitions was the race window. A SuspendNotifications()
landing in it would see a zero count and skip emitting 'true', and the resume would
then emit 'false' on top of it, leaving the count and the subject in disagreement.

Emitting the signal inside the same scope that cleared the count removes the window
entirely, rather than detecting and compensating for it after the fact.

Delivery is still performed off the lock: ScopedAccess.Dispose() calls
ExitLockAndDeliver(), which releases the lock before draining the queue. The signal
now fires before the pending changes are delivered instead of after, which a deferred
subscriber handles via the existing HasPending/_currentDeliveryVersion guard in
CreateConnectObservable.

Also guard EmitResumeNotification() against a disposed subject, since
SuspensionTracker.Dispose() runs from the delivery callbacks off the lock and can
tear the subject down underneath a concurrent resume.

With the window closed, ResumeSignalUnderLockPreventsStaleSnapshotFromReSuspend is
restored and passes alongside StaleResumeSignalIsSuppressedByConcurrentReSuspend.

---------


(cherry picked from commit 4f0b4573ec4b0d8e854778bf1b60aa26f3d1c1cb)
